renamed the button components
